requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
accelerate=0.22.0=pypi_0
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
anyio=4.0.0=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.2.3=pypi_0
asttokens=2.4.0=pypi_0
async-lru=2.0.4=pypi_0
async-timeout=4.0.3=pypi_0
attrs=23.1.0=pypi_0
babel=2.12.1=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
bitsandbytes=0.41.1=pypi_0
black=23.9.1=pypi_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.08.22=haa95532_0
certifi=2023.7.22=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=3.2.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.1.4=pypi_0
datasets=2.14.5=pypi_0
debugpy=1.8.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dill=0.3.7=pypi_0
docker-pycreds=0.4.0=pypi_0
exceptiongroup=1.1.3=pypi_0
executing=1.2.0=pypi_0
fastjsonschema=2.18.0=pypi_0
filelock=3.12.4=pypi_0
fqdn=1.5.1=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.6.0=pypi_0
gitdb=4.0.10=pypi_0
gitpython=3.1.36=pypi_0
huggingface-hub=0.17.1=pypi_0
icc_rt=2022.1.0=h6049295_2
idna=3.4=pypi_0
intel-openmp=2023.1.0=h59b6b97_46319
ipykernel=6.25.2=pypi_0
ipython=8.15.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=8.1.1=pypi_0
isoduration=20.11.0=pypi_0
isort=5.12.0=pypi_0
jedi=0.19.0=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=py310haa95532_0
json5=0.9.14=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.19.0=pypi_0
jsonschema-specifications=2023.7.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.3.1=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.3.1=pypi_0
jupyter-events=0.7.0=pypi_0
jupyter-lsp=2.2.0=pypi_0
jupyter-server=2.7.3=pypi_0
jupyter-server-terminals=0.4.4=pypi_0
jupyterlab=4.0.5=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-server=2.25.0=pypi_0
jupyterlab-widgets=3.0.9=pypi_0
libffi=3.4.4=hd77b12b_0
markupsafe=2.1.3=pypi_0
matplotlib-inline=0.1.6=pypi_0
mistune=3.0.1=pypi_0
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py310h2bbff1b_1
mkl_fft=1.3.6=py310h4ed8f06_1
mkl_random=1.2.2=py310h4ed8f06_1
multidict=6.0.4=pypi_0
multiprocess=0.70.15=pypi_0
mypy-extensions=1.0.0=pypi_0
nbclient=0.8.0=pypi_0
nbconvert=7.8.0=pypi_0
nbformat=5.9.2=pypi_0
nest-asyncio=1.5.7=pypi_0
notebook=7.0.3=pypi_0
notebook-shim=0.2.3=pypi_0
numpy=1.25.2=py310h055cbcc_0
numpy-base=1.25.2=py310h65a83cf_0
openssl=1.1.1w=h2bbff1b_0
overrides=7.4.0=pypi_0
packaging=23.1=pypi_0
pandas=2.1.0=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pathspec=0.11.2=pypi_0
pathtools=0.1.2=pypi_0
peft=0.5.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.0.0=pypi_0
pip=23.2.1=py310haa95532_0
platformdirs=3.10.0=pypi_0
prometheus-client=0.17.1=pypi_0
prompt-toolkit=3.0.39=pypi_0
protobuf=4.24.3=pypi_0
psutil=5.9.5=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=13.0.0=pypi_0
pycparser=2.21=pypi_0
pygments=2.16.1=pypi_0
python=3.10.9=h966fe2a_2
python-dateutil=2.8.2=pypi_0
python-json-logger=2.0.7=pypi_0
pytz=2023.3.post1=pypi_0
pywin32=306=pypi_0
pywinpty=2.0.11=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.1=pypi_0
qtconsole=5.4.4=pypi_0
qtpy=2.4.0=pypi_0
referencing=0.30.2=pypi_0
regex=2023.8.8=pypi_0
requests=2.31.0=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.10.3=pypi_0
safetensors=0.3.3=pypi_0
scikit-learn=1.3.0=py310h4ed8f06_0
scipy=1.11.2=pypi_0
send2trash=1.8.2=pypi_0
sentry-sdk=1.31.0=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=68.0.0=py310haa95532_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h2bbff1b_0
stack-data=0.6.2=pypi_0
tbb=2021.8.0=h59b6b97_0
terminado=0.17.1=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tinycss2=1.2.1=pypi_0
tk=8.6.12=h2bbff1b_0
tokenizers=0.13.3=pypi_0
tomli=2.0.1=pypi_0
torch=1.13.1+cu116=pypi_0
torchaudio=0.13.1+cu116=pypi_0
torchvision=0.14.1+cu116=pypi_0
tornado=6.3.3=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.9.0=pypi_0
transformers=4.33.1=pypi_0
typing-extensions=4.7.1=pypi_0
tzdata=2023.3=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.0.4=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wandb=0.15.10=pypi_0
wcwidth=0.2.6=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.6.3=pypi_0
wheel=0.38.4=py310haa95532_0
widgetsnbextension=4.0.9=pypi_0
xxhash=3.3.0=pypi_0
xz=5.4.2=h8cc25b3_0
yarl=1.9.2=pypi_0
zlib=1.2.13=h8cc25b3_0
